+++ include/tasks.h
@@ -18,29 +18,29 @@ public:
     ///element
     Task *data();
     const Task *data() const;
-    uint64_t size() const;
-    Task &operator[](uint64_t taskID);
-    Task &at(uint64_t taskID);
+    size_t size() const;
+    Task &operator[](size_t taskID);
+    Task &at(size_t taskID);
     ///select
     //general
-    std::vector<uint64_t> select(std::function<bool(const Task &)> pred) const;
+    std::vector<size_t> select(std::function<bool(const Task &)> pred) const;
     //equal
-    std::vector<uint64_t> select(const Task &) const;
+    std::vector<size_t> select(const Task &) const;
     //member
     template<class Member>
-    std::vector<uint64_t> select(Member Task:: *pointer, Member value) const;
+    std::vector<size_t> select(Member Task:: *pointer, Member value) const;
     ///sort
     //general
-    std::vector<uint64_t> sort(std::function<bool(const Task &, const Task &)> cmp) const;
+    std::vector<size_t> sort(std::function<bool(const Task &, const Task &)> cmp) const;
     //member
     template<class Member>
-    std::vector<uint64_t> sort(Member Task:: *pointer, 
+    std::vector<size_t> sort(Member Task:: *pointer, 
         std::function<bool(const Member &, const Member &)> cmp = std::less<Member>()) const;
 };
 
 //inline
 template<class Member>
-inline std::vector<uint64_t> Tasks::select(Member Task:: *pointer, Member value) const
+inline std::vector<size_t> Tasks::select(Member Task:: *pointer, Member value) const
 {
     return select([&](const Task &x) -> bool
         {
@@ -49,7 +49,7 @@ inline std::vector<uint64_t> Tasks::select(Member Task:: *pointer, Member value)
 }
 
 template<class Member>
-inline std::vector<uint64_t> Tasks::sort(Member Task:: *pointer, 
+inline std::vector<size_t> Tasks::sort(Member Task:: *pointer, 
     std::function<bool(const Member &, const Member &)> cmp) const
 {
     return sort([&](const Task &x, const Task &y)
